+++ cmd/sidecar-injector/main.go
@@ -20,9 +20,9 @@ func main() {
 	flag.IntVar(&parameters.Port, "port", 443, "Webhook server port.")
 	flag.StringVar(&parameters.CertFile, "tlsCertFile", "/etc/webhook/certs/cert.pem", "File containing the x509 Certificate for HTTPS.")
 	flag.StringVar(&parameters.KeyFile, "tlsKeyFile", "/etc/webhook/certs/key.pem", "File containing the x509 private key to --tlsCertFile.")
+	flag.BoolVar(&parameters.NoTLS, "noTLS", false, "Disable SSL and ignore any certs.")
 	flag.StringVar(&parameters.SecretlessContainerImage, "secretless-image", "cyberark/secretless-broker:latest", "Container image for the Secretless sidecar")
 	flag.StringVar(&parameters.AuthenticatorContainerImage, "authenticator-image", "cyberark/conjur-kubernetes-authenticator:latest", "Container image for the Kubernetes Authenticator sidecar")
-	flag.BoolVar(&parameters.NoTLS, "noTLS", false, "Disable SSL and ignore any certs.")
 	flag.Parse()
 
 	whsvr := &inject.WebhookServer{