+++ pkg/inject/patch.go
@@ -21,13 +21,43 @@ const (
 )
 
 // create mutation patch for resources
-func createPatch(pod *corev1.Pod, sidecarConfig *PatchConfig, annotations map[string]string) ([]byte, error) {
+func createPatch(
+	pod *corev1.Pod,
+	sidecarConfig *PatchConfig,
+	annotations map[string]string,
+) ([]byte, error) {
 	var patch []rfc6902PatchOperation
 
-	patch = append(patch, addContainer(pod.Spec.InitContainers, sidecarConfig.InitContainers, "/spec/initContainers")...)
-	patch = append(patch, addContainer(pod.Spec.Containers, sidecarConfig.Containers, "/spec/containers")...)
-	patch = append(patch, addVolume(pod.Spec.Volumes, sidecarConfig.Volumes, "/spec/volumes")...)
-	patch = append(patch, updateAnnotation(pod.Annotations, annotations)...)
+	patch = append(
+		patch,
+		addContainer(
+			pod.Spec.InitContainers,
+			sidecarConfig.InitContainers,
+			"/spec/initContainers",
+		)...,
+	)
+	patch = append(
+		patch,
+		addContainer(
+			pod.Spec.Containers,
+			sidecarConfig.Containers,
+			"/spec/containers",
+		)...,
+	)
+	patch = append(
+		patch,
+		addVolume(
+			pod.Spec.Volumes,
+			sidecarConfig.Volumes, "/spec/volumes",
+		)...,
+	)
+	patch = append(
+		patch,
+		updateAnnotation(
+			pod.Annotations,
+			annotations,
+		)...,
+	)
 	patch = append(
 		patch,
 		addVolumeMounts(
@@ -41,9 +71,13 @@ func createPatch(pod *corev1.Pod, sidecarConfig *PatchConfig, annotations map[st
 }
 
 // addContainer create a patch for adding containers
-func addContainer(target, added []corev1.Container, basePath string) (patch []rfc6902PatchOperation) {
+func addContainer(
+	target, added []corev1.Container,
+	basePath string,
+) (patch []rfc6902PatchOperation) {
 	first := len(target) == 0
 	var value interface{}
+
 	for _, add := range added {
 		value = add
 		path := basePath
@@ -59,14 +93,21 @@ func addContainer(target, added []corev1.Container, basePath string) (patch []rf
 			Value: value,
 		})
 	}
+
 	return patch
 }
 
 // addVolumeMounts creates a patch for adding volume mounts
-func addVolumeMounts(target []corev1.Container, added ContainerVolumeMounts, basePath string) (patch []rfc6902PatchOperation) {
+func addVolumeMounts(
+	target []corev1.Container,
+	added ContainerVolumeMounts,
+	basePath string,
+) (patch []rfc6902PatchOperation) {
 	for index, container := range target {
 		volumeMounts, ok := added[container.Name]
-		if !ok || len(volumeMounts) == 0 { continue }
+		if !ok || len(volumeMounts) == 0 {
+			continue
+		}
 
 		if len(container.VolumeMounts) == 0 {
 			volumeMount := volumeMounts[0]
@@ -89,36 +130,47 @@ func addVolumeMounts(target []corev1.Container, added ContainerVolumeMounts, bas
 			})
 		}
 	}
+
 	return patch
 }
 
 // addVolume creates a patch for adding volumes
-func addVolume(target, added []corev1.Volume, basePath string) (patch []rfc6902PatchOperation) {
+func addVolume(
+	target, added []corev1.Volume,
+	basePath string,
+) (patch []rfc6902PatchOperation) {
 	first := len(target) == 0
 	var value interface{}
+
 	for _, add := range added {
 		value = add
 		path := basePath
+
 		if first {
 			first = false
 			value = []corev1.Volume{add}
 		} else {
 			path = path + "/-"
 		}
+
 		patch = append(patch, rfc6902PatchOperation{
 			Op:    patchOperationAdd,
 			Path:  path,
 			Value: value,
 		})
 	}
+
 	return patch
 }
 
 // updateAnnotation creates a patch for adding/updating annotations
-func updateAnnotation(target, added map[string]string) (patch []rfc6902PatchOperation) {
+func updateAnnotation(
+	target, added map[string]string,
+) (patch []rfc6902PatchOperation) {
 	for key, value := range added {
 		if target == nil || target[key] == "" {
 			target = map[string]string{}
+
 			patch = append(patch, rfc6902PatchOperation{
 				Op:   patchOperationAdd,
 				Path: "/metadata/annotations",
@@ -134,5 +186,6 @@ func updateAnnotation(target, added map[string]string) (patch []rfc6902PatchOper
 			})
 		}
 	}
+
 	return patch
 }